move everything related to shapefile drawing into separate module
